Add files via upload
The installation process for PHP is a relatively simple process on the Windows and the Linux operating system. The first step begins with the download of the file “php-8.1.12.tar.gz” from “php.net”. Once the tar.gz file is in the download folder. The next step involves the use of terminal. To install type the following command: 
Windows PHP installation command: 
sudo apt install php
Terminal will ask you if you want to proceed with the installation. Type ‘y’ to choose yes as your response. Upon successful installation, you will be able to verify where PHP was installed by typing “what php”. You also have the option to install a PHP add-on in Visual Studio Code [2]. 

III. PHP Installation Mac
In the MacOS operating system, the installation of the PHP language requires the use of the use of the program Homebrew to install the “php-8.1.12.tar.gz” file. The command highlighted by the red box in the image below must be copied into terminal to install Homebrew. 
 
Homebrew installation command: 
/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
Once Homebrew has been installed, the next step is to type in the command below to install PHP with the .tar.gz file [2]. 
Mac PHP installation command:
brew install php